requirements.txt
alembic==1.18.1
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
bcrypt==3.2.0
billiard==4.2.4
boto3==1.42.33
botocore==1.42.33
celery==5.6.2
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cryptography==46.0.3
distro==1.9.0
dnspython==2.8.0
dotenv==0.9.9
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.128.0
greenlet==3.3.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
jiter==0.12.0
jmespath==1.1.0
kombu==5.6.2
Mako==1.3.10
mangum==0.20.0
MarkupSafe==3.0.3
openai==2.15.0
packaging==26.0
passlib==1.7.4
pip==24.3.1
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
pyasn1==0.6.2
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.21
pytz==2025.2
PyYAML==6.0.3
redis==7.1.0
requests==2.32.5
rsa==4.9.1
s3transfer==0.16.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.45
starlette==0.50.0
tqdm==4.67.1
typing_extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
uvicorn==0.40.0
vine==5.1.0
watchfiles==1.1.1
wcwidth==0.3.1
websockets==16.0
rapidfuzz==3.12.1
pgvector==0.3.6
langchain-openai==0.3.3
langchain-text-splitters==0.3.6
pypdf==4.0.1
langchain-google-genai==0.0.9
